Add HTML semantic tags, move media query to the end and apply seperation of concerns to the JS
