+++ src/index.html
@@ -7,7 +7,8 @@
     <title>To Do List</title>
 </head>
 <body>
-    <div id="toDoForm">
+    <header></header>
+    <main id="toDoForm">
         <div class="title">
             <h1>Today's To Do</h1>
             <i class="fa fa-arrows-rotate icon"></i>
@@ -20,7 +21,8 @@
             </form>
         <div class="toDoTasksContainer"></div>
         <button class="btn_clear_all" id="btnClear" type="button">Clear all completed</button>
-    </div>
+    </main>
+    <footer></footer>
     <script src="../dist/main.js"></script>
 </body>
 </html>
